+++ src/main/java/hr/dlatecki/algorithms/gen_alg/operators/mutation/BitFlipMutation.java
@@ -4,22 +4,22 @@ package hr.dlatecki.algorithms.gen_alg.operators.mutation;
 
 import java.util.Random;
 import hr.dlatecki.algorithms.gen_alg.operators.abstracts.AbstractMutationOperator;
-import hr.dlatecki.algorithms.gen_alg.population.chromosomes.ByteArrayChromosome;
+import hr.dlatecki.algorithms.gen_alg.population.abstracts.AbstractByteArrayChromosome;
 
 /**
  * An implementation of mutation operator which flips bits in order to mutate the chromosome. Bits are flipped with
  * provided probability rate.<br>
  * <br>
- * This mutation operator can only be used on chromosomes which extend the <code>ByteArrayChromosome</code>.
+ * This mutation operator can only be used on chromosomes which extend the <code>AbstractByteArrayChromosome</code>.
  * 
  * @author Domagoj Latečki
  * @version 1.0
  * @since 1.8
  * @param <C> the type of chromosome which will be used in mutation operator.
  * @see AbstractMutationOperator
- * @see ByteArrayChromosome
+ * @see AbstractByteArrayChromosome
  */
-public class BitFlipMutation<C extends ByteArrayChromosome> extends AbstractMutationOperator<C> {
+public class BitFlipMutation<C extends AbstractByteArrayChromosome> extends AbstractMutationOperator<C> {
     
     /**
      * Constructs a bit flip mutation operator. Mutation intensity specifies the chance for each bit to be flipped.