+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Development history
 
 ### v0.6.1
-Removed annotations package and all its contents.
+Removed annotations package and all its contents. Added MIN_SELECTION_SIZE constant in ISelection operator interface.
+All chromosome implementations are now marked as final.
 
 ### v0.6.0
 Added ArithmeticMeanCrossover operator.